API-14:  add fidor oauth+fidor api url to examples. update starter kit builder to parse the new values, update readme

+++ lib/fidor_starter_kits.rb
@@ -7,23 +7,31 @@ module FidorStarterKits
 
   class << self
 
-    # Path to starter kits
+    # Directory path to starter kits
     def path
       File.expand_path('../starter_kits', File.dirname(__FILE__))
     end
 
+    # Check if an app with the given name exists in the starter kits directory
+    # @param [String] app_name
+    # @return [Boolean]
     def exists?(app_name)
       File.exists? File.join(path, app_name)
     end
 
-    # @param [String] app_name  directory name of a starter kit
-    # @param [String] client_id
-    # @param [String] client_secret
-    # @param [String] app_url
-    # @param [String] fidor_url
-    # @return [String] path to zipped example im /tmp folder
-    def build(app_name, client_id, client_secret, app_url, fidor_url)
-      # move example to a save location
+    # @param [Hash] opts options for building a starter kit
+    # @options opts [String] :app_name directory name of a starter kit
+    # @options opts [String] :client_id app client id
+    # @options opts [String] :client_secret app client secret
+    # @options opts [String] :app_url full url to the application
+    # @options opts [String] :fidor_oauth_url full url to the oauth endpoints e.g http://fidor.de/oauth
+    # @options opts [String] :fidor_api_url full url to the api or api_sandbox no trailing slashes e.g http://fidor.de/api_sandbox
+    # @return [Nil | String] path to zipped example im /tmp folder or nil if app does not exists
+    def build(opts)
+      app_name = opts[:app_name]
+      return if !app_name || !exists?(app_name)
+
+      # move example to a safe location
       example_src_path = File.join(path, app_name)
       tmp_src_dir = Dir.mktmpdir(app_name)
       FileUtils.copy_entry example_src_path, tmp_src_dir
@@ -32,11 +40,10 @@ module FidorStarterKits
       example_files =  File.join(tmp_src_dir, 'example.*')
       Dir.glob(example_files) do |example_file|
         content = File.read(example_file)
-        changed_content = content.gsub(/<CLIENT-ID>/, client_id)
-                                  .gsub(/<CLIENT-SECRET>/, client_secret)
-                                  .gsub(/<APP-URL>/, app_url)
-                                  .gsub(/<FIDOR-URL>/, fidor_url)
-        File.write(example_file, changed_content)
+        %w(client_id client_secret app_url fidor_oauth_url fidor_api_url).each do |field|
+          content.gsub!("<#{field.upcase}>", opts[field.to_sym]) if opts[field.to_sym]
+        end
+        File.write(example_file, content)
       end
 
       # create zip file in tmp dir
